vfio-pci: add a switch to disable PCI AER
AER is meant to let a device driver recover from
errors discovered on the PCIe bus. However, the
current implementation of vfio-pci does not distingish
between correctable or uncorrectalbe as well as fatal
vs. non-fatal errors. Any kind of error can trigger the
error correction interrupt and cause all vServers to
switch to RUN_STATE_INTERNAL_ERROR.

I have observed correctable non-fatal errors on a
PCI root hub which where then propagated to all vServers
on this root hub causing them to shut down.

I added this switch to be able to ignore AER interrupts
until a proper interface to propagate the error type from
kernel to qemu is there. That would be the old behaviour
of pci-assign or a host kernel not supporting AER.

Signed-off-by: Peter Lieven <pl@kamp.de>

Conflicts:
	hw/vfio/pci.c
	hw/vfio/pci.h

Conflicts:
	hw/vfio/pci.h